+++ CHANGELOG.md
@@ -6,7 +6,8 @@
     * Added `LogViewer` helper class and CLI command: `bin/grav logviewer`
 1. [](#improved)
     * Improved error detection for broken Flex Objects
-    * Removed apc and xcache support, made apc alias of apcu
+    * Removed `apc` and `xcache` support, made `apc` alias of `apcu`
+    * Support admin and regular translations via the `|t` twig filter and `t()` twig function
 1. [](#bugfix)
     * Fixed non-namespaced exceptions in scheduler
     * Fixed trailing slash redirect in multlang environment [#2350](https://github.com/getgrav/grav/issues/2350)